requirements.txt
pyttsx3
pyyaml
SpeechRecognition
